iclean/requirements.txt
amqp==5.1.1
asgiref==3.5.1
atomicwrites==1.4.0
attrs==21.4.0
autopep8==1.6.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
commonmark==0.9.1
coreapi==2.3.3
coreschema==0.0.4
coverage==6.4.1
cryptography==37.0.2
defusedxml==0.7.1
Django==4.0.4
django-celery-beat==2.3.0
django-celery-results==2.3.1
django-debug-toolbar==3.4.0
django-extensions==3.1.5
django-filter==21.1
django-templated-mail==1.1.1
django-timezone-field==5.0
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-yasg==1.20.0
factory-boy==3.2.1
Faker==13.12.0
flower==1.0.0
gevent==21.12.0
greenlet==1.1.2
httpie==3.2.1
humanize==4.1.0
idna==3.3
inflection==0.5.1
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
model-bakery==1.5.0
multidict==6.0.2
oauthlib==3.2.0
openapi-codec==1.3.2
packaging==21.3
pluggy==1.0.0
prometheus-client==0.14.1
prompt-toolkit==3.0.29
py==1.11.0
pycodestyle==2.8.0
pycparser==2.21
Pygments==2.12.0
PyJWT==2.4.0
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.1.2
pytest-django==4.5.2
pytest-factoryboy==2.4.0
python-crontab==2.6.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rich==12.4.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
tornado==6.1
typing_extensions==4.2.0
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
zope.event==4.5.0
zope.interface==5.4.0
